Version 0.1.35
* Added shell scripts for deployment from from scratch included Node.js,
MongoDB and Impress
- For CentOS /deploy/centos.sh tested for CentOS 6.5 64bit minimal (may
work for Red Hat but not tested yet)
- For Ubuntu /deploy/ubuntu.sh tested for Ubuntu 14.04 64bit minimal
- For Debian /deploy/debian.sh tested for Debian 7.5 64bit minimal
* Fixed /bin/install.js and /bin/uninstall.js for ubuntu and debian

+++ bin/uninstall.sh
@@ -1,5 +1,9 @@
 #!/bin/sh
 service impress stop
-chkconfig impress off
-chkconfig --del impress
-rm /etc/init.d/impress
+if [ -f /etc/debian_version ]; then
+	sudo update-rc.d impress disable
+elif [ -f /etc/redhat-release ]; then
+	chkconfig impress off
+	chkconfig --del impress
+	rm /etc/init.d/impress
+fi
